Handle
New module under desper.data, used for resource management. A handle encapsulates the information needed to generate/load a resource but doesn't actually loads it until requested the first time(the resource is then cached for further use).

Handle is a base class that can(and will) be specialized to obtain some dynamic resource loading when requested.