GameModel as a resource controller(WIP)
Added the gamemodel internal module, containing the GameModel class.
This is a base class used as a resource and main game logic controller.
So far, it's capable of importing resources using specifically designed
importer dictionaries. It imports Handles in the GameModel.res dict.

The exact signature is GameModel(dirs=[], importer_dict={})

example dir structure:
 resources
 |_res1.txt
 |_res2.txt

Giving ['resources'] as parameter to the GameModel, the resources will
be loaded as follows:
 # ...

The importer_dict parameter helps you specify which specific Handle type
should be applied to which resource, using the format:
{function1: HandleType1, function2: HandleType2, ...}

function1 should accept a filepath and should return an iterable of
parameters to be passed to the HandleType constructor in order to
initialize it(if the resource is considered accepted by that function).
The function should return None if the resource is not to be considered
accepted by that type of Handle.

A resource which mathces with no function will not be loaded into the
dictionary. A resource which matches with more than one function will be
loaded by more than one Handle(this might be what you want, but probably
not, so be carefull if you define your own importer_dicts' functions and
HandleTypes).

The importer_dict paramenter