ImageHandle for pyglet.image.AbstractImage
The ImageHandle will load a pyglet.image.AbstractImage using
pyglet.resource.image. This means that it's a Handle specifically
deisgned for static images(for pyglet.image.Animation a specific Handle
will be implemented).

With ImageHandle, get_image_importer() has been implemented. This function will return an importer function suitable as key in an importer_dict(for GameModels).
It accepts two optional parameters, location and accepted_exts, which
will be injected in the resulting funciton thanks to the environment.