Glet module AnimationHandle
An animation handle accepts a filename pointing to a json file, and generates a pyglet.image.Animation from metadata read from that file.

The json data structure needed will be documented in the near future, but it's a subset of the exported data for animations from aseprite(https://www.aseprite.org/), to keep it as easy to generate as possible.