No extension mode tweaks
Made some improvements for scalability reasons:
The "no extensions" mode for resources is now specified in a global set of options(actually, it's the only existing options so far), which can be found in the core module desper.options(it's a dictionary called options). The module is self imported when importing desper.