Added OnAttachListener interface
The interface can be used with components(AbstractComponents, but not necessarily) and must be used with AbstractWorlds.

The interface implements an on_attach(entity, world) method, which will be called whenever a component implementing it is adyded to the world(through entity_create or add_component).

This can be used for initialization of components(like components dynamically creating more components, or caching of other components from the current world).

Unfortunately, the current GameModel instance can't be directly passed to the on_attach method, meaning that if the model is requested, it has to be acquired in other ways).