Added Camera and CameraProcessor
Camera is an ECS component defining a camera(for now, position, viewport and zoom are supported).
CameraProcessor is the associated ECS processor that renders everything according to Camera components in the current World. No rendering is done anymore by the GletGameModel(yet, it cleans the screen), so if you want to render something you have to instantiate at least a Camera and a CameraProcessor).

Warning: CameraProcessor assumes the use of a GletGameModel(since for practical reasons, the model contains the pyglet batches for each world).
This is slightly against the general design of the library. but greatly increases usability(yet it's not catastrophic since desper.glet ECS processors/components are meant to be used with a GletGameModel anyway, even if they could work on their own with just a World).