Added EventHandler decorator class
Instances of EventHandler are callables, and should be used as decorators. When used as decorator for classes. The decorated class will be automatically added by EventHandler to the pyglet.window handlers(using pyglet.window.Window.push_handlers).

This also means that each instance of EventHandler manages one window. A default instance is constructed at desper.glet.event_handler (but it creates no window by default, so it needs to be manually set using the EventHandler.window attribute(this is required ofc for any newly constructed EventHandler too).