Better on_attach event logic
The on_attach event can now be prevented from being triggered, thanks to proper arguments in the AbstractWorld methods.

Also, when creating a new entity with AbstractWorld.create_entity, on_attach will be called only after all the components have been added. This means that (eventual) component retrieval inside the on_attach method doesn't depend on the order in which the components are given to the method.