Added coroutines
A coroutine in desper, is simply a generator(important: it's NOT a
python coroutine object).
Generators are the perfect solution, since the execution is arbitrated by
an ECS processor(CoroutineProcessor), and they keep the state of the
execution thanks to yield.

When yielding a number(an integer: n), the coroutine will be paused for
n frames(while paused, it will have almost no resource usage).