CoroutineState enum for coroutine queries
It's now possible to query a CoroutineProcessor for the current state of a specific coroutine.
The possible results are: TERMINATED(the coroutine isn't found in the system: it's dead), PAUSED(the coroutined yielded and it's waiting for some frames), ACTIVE(the coroutine is active and will be processed as soon as possible).