Added component_initializers dict to WorldHandle
With component_initializers it's possible to define custom initializer
functions for specific component types. The default initializer uses the
given type(which assumes to be a of type "type", a class basically) to
initialize the component, passing the given arguments from the
json(args, kwargs).
If a specific initializer for a type is given, the default behaviour is
overwritten.