Added glet.ActivePositionProcessor
A general processor used to update components based on glet.Position. It accepts the type of the components to update (it's pure ECS, so no subtyping) along with some customization.

To update sprites, ActiveSpriteProcessor is kept since it's a bit more efficient. ActivePositionProcessor can be used too, at the cost of efficiency.