Better glet.Camera management
It is now possible to specify both the projection and the viewport. The projection defaults to window size, and the viewport defaults to the best fit in the window that keeps projection's aspect ratio.

If a viewport is specified, the projection will freely stretch to fit it (no ratio involved).