CoroutineProcessor: factory for delta time
Previously, time in for coroutines was interpreted in frames. Now,
through the use of a factory it can be interpreted as seconds (or
fractions) too.

The default factory is simply a lamda: 1, keeping the count on frames.
By replacing it with a custom factory similar to this:
  lambda: model.dt

A custom delta time can be injected, making the time count for
coroutines in real time seconds.