Added ResourceMap.get_static_map()
get_static_map() generates and returns a StaticResourceMap that mirrors the hierarchy of the original ResourceMap.
Access is much more slim (directly through attribute lookup) and efficient than through the original ResourceMap. Modifications to the original map do not propagate to the static one/s. Static maps are immutable (no new items can be added or removed, an entirely new static map shall be generated in case of edits).