Fixed removal/addition issues during ev. dispatch
Removing or adding handlers during the dispatching of an event could cause a runtime error since the size of the inspected set could change during iteration.
This was fixed by modifying dispatch(...) so that the set of handlers for the interested event is copied before iterating it.
This is an unfortunate slowdown, but also the simplest and most reliable way to prevent such conflicts.
Similar but different strategies may be applied in the future based on benchmarks, trying to optimize this behaviour.

Adding a check on the existance of the referents may be necessary in the future.