Wolrd.add_component, on_add w/ dispatch disabled
When dispatching is disabled, on_add shall be enqueued as well. Due to the locality of this event (event system assumes all events are global scoped with respect to the dispatcher) it is not possible to dispatch it with the given dispatch(...) method. Instead, World listens and dispatches a relay event, that is correctly enqueued and triggers on_add.

To gain some performance when adding components, if event dispatching is enabled on_add is immediately called (passing through the relay event would be possible, but unnecessarily slow).