Added World.get(...), query all components on type
All components of the given type are returned. Actually, all pairs (entity, component) where entity owns component and component has the given type (or is a subinstance of it).
An internal generator method _get(...) actually does the query. The exernal interface simply wraps _get(...) and returns a list containing all the elements. This is because in the near future get(...) will keep a cache of the queries.