Changed on_add_dispatch in on_component_dispatch
Renamed World's self-handled event used to dispatch to single components. Moreover, one more parameter is now required. That is, the event to be dispatched. In practice, the on_add_dispatch has now become a general way to dispatch events to single handlers, and that is why it has been renamed.