World._on_component_dispatch-> _on_single_dispatch
Renamed and refactored the event to make it more general. Dispatching can now be done towards any individual event handler. This will be useful when dealing with processors, which can be event handlers but are not components (eg. no entity id to dispatch, and no need to dispatch the world).