Added World.clear()
Clear the whole database. Entities are deleted, then processors are removed. Finally, the event dispatching system is cleared as well.

The id generator is not being reset. Slight changes to the interface are necessary to do so. In particular, a factory has to be provided during creation of the World (conversely to the current interface that simply accepts a generator).