Fixed on_add dispatch issue w/ dispatch disabled
When dispatching is disabled, World enqueues a single dispatch event to itself. This event functions as a relay, which calls on_add when dispatching is enabled. A small logical bug caused the single dispatch event to be enqueued even if the inserted handler was not handling the on_add event.