WiP World loading from file
Through dictionary manipulation. A well formed dictionary is then
submitted to populate_world_with_dict(...) which finally instantiates
all the components.

Dictionary manipulators are called "dictiornary transformers". The
behaviour is similar to the one of standard callable transformers, but
act on a "lower" level. Instead of modifying the world directly, they
modify the dictionary that will populate it.

All this will be pulled off by using the new classes WorldFromFileHandle
and WorldFromFileTransformer. The transformer is working and tested.

Default dictionary transformers shall be defined, so that the
WorldFromFileTransformer can be nimbly applied to a project.